### 2022-02-02 #4126 @whedon
[PRE REVIEW]: SimplyAgree: An R package and jamovi Module for Simplifying Agreement and Reliability Analyses
**Submitting author:** @arcaldwell49 (<a href="http://orcid.org/0000-0002-4541-6283">Aaron R. Caldwell</a>)
**Repository:** <a href="https://github.com/arcaldwell49/SimplyAgree" target ="_blank">https://github.com/arcaldwell49/SimplyAgree</a>
**Version:** v0.0.2
