+++ DESCRIPTION
@@ -7,7 +7,7 @@ Authors@R: person("Aaron", "Caldwell", email = "arcaldwell49@gmail.com",
 Maintainer: Aaron Caldwell <arcaldwell49@gmail.com>
 Description: Reliability and agreement analyses often have limited software support. Therefore, this package was created to make agreement and reliability analyses easier for the average researcher. The functions within this package include simple tests of agreement, agreement analysis for nested and replicate data, and provide robust analyses of reliability. In addition, this package contains a set of functions to help when planning studies looking to assess measurement agreement. For robust analyses of agreement, limits of agreement through a bootstrap method can also be calculated.
 URL: https://aaroncaldwell.us/SimplyAgree/
-License: GPL-3
+License: GPL (>= 3)
 Encoding: UTF-8
 LazyData: true
 RoxygenNote: 7.1.1
