### 2022-02-02 #4126 @arcaldwell49
My sincerest apologies, I am pushing the license file to GitHub now. Also,
just as a note, one article failed the DOI check. I am unsure what the
issue is here but this is the DOI listed on the publication and on PubMed.

On Wed, Feb 2, 2022 at 8:54 AM Kevin Mattheus Moerman <
***@***.***> wrote:

> @arcaldwell49 <https://github.com/arcaldwell49> our system could not see
> an open source license. Can you add a plain LICENSE file clarifying the
> license in the main repository folder? Note this needs to be an Open
> Source Initiative approved license <https://opensource.org/licenses>.
>
