### 2022-02-02 #4126 @Kevin-Mattheus-Moerman
@arcaldwell49 The DOI is missing for this one, please can you add it?

```bib
@article{weir2005,
  url = {https://pubmed.ncbi.nlm.nih.gov/15705040/},
  year = {2005},
  volume = {19},
  number = {1},
  pages = {231},
  author = {Joseph P. Weir},
  title = {Quantifying Test-Retest Reliability Using the Intraclass Correlation Coefficient and the {SEM}},
  journal = {The Journal of Strength and Conditioning Research}
}
```

I think adding `doi = {10.1519/15184.1}` would do it.
