### 2022-02-02 #4126 @Kevin-Mattheus-Moerman
@fboehm you have been suggested as editor for this work. Do you think it is a fit for you? The paper doesn't compile yet but the raw form is here: https://raw.githubusercontent.com/arcaldwell49/SimplyAgree/master/paper/paper.md
