### 2022-02-02 #4126 @Kevin-Mattheus-Moerman
@arcaldwell49 I see. I checked just now too, so it looks like there is no DOI. In that case we'll ignore this warning message and you can leave the bib file as is.
