### 2022-02-10 #4148 @whedon
[REVIEW]: SimplyAgree: An R package and jamovi Module for Simplifying Agreement and Reliability Analyses
**Submitting author:** @arcaldwell49 (<a href="http://orcid.org/0000-0002-4541-6283">Aaron R. Caldwell</a>)
**Repository:** https://github.com/arcaldwell49/SimplyAgree
**Branch with paper.md** (empty if default branch):
**Version:** v0.0.3

- [x] #4214
