### 2022-02-24 #4126 @arcaldwell49
All, I just want to give a small update. A user of the package noticed an error under certain conditions for agree_nest and agree_reps. I have put in a small patch (v0.0,3) to fix the issue. I am merging a branch with the main branch now and will submit to CRAN soon.
