### 2022-02-25 #4148 @slwu89
Hi @fboehm, how do I regenerate the PDF? It looks like the old whedon command doesn't work. Thanks!
