### 2022-02-25 9aeb034
Update .gitignore
Ignore mac DS_store files
