### 2022-02-25 arcaldwell49/SimplyAgree#9 @slwu89
vignette/docs comments
1. In line https://github.com/arcaldwell49/SimplyAgree/blob/9aeb034a7b3dbfaefa6ef69e154dba9360b51880/vignettes/intro_vignette.Rmd#L71 \gamma_{1-\alpha} is not surrounded by `$` so it does not render on the docs site.
2. I suggest adding a link to the docs website (https://aaroncaldwell.us/SimplyAgree/index.html) somewhere to the repo main page since some users won't read the JOSS paper to find it, either on the README.md or in the sidebar "About" to the right, under "website". It can also be added to the `URL` field of the DESCRIPTION (https://cran.r-project.org/doc/manuals/R-exts.html#The-DESCRIPTION-file)
