### 2022-02-25 arcaldwell49/SimplyAgree#10 @slwu89
community guidelines
According to JOSS requirements, the repo documentation should include clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support.

I think to address that, some things to do are:

- Add the GitHub issues link to `BugReports` in DESCRIPTION (as in https://cran.r-project.org/doc/manuals/R-exts.html#The-DESCRIPTION-file)
- Include a section in README.md, "Contributing" indicating how to file a github issue and how PRs or other contributions or questions should be handled.
