### 2022-02-25 #4148 @slwu89
I am linking issues created in the review repo to this review here:

- community guidelines https://github.com/arcaldwell49/SimplyAgree/issues/10
- finding documentation https://github.com/arcaldwell49/SimplyAgree/issues/9
