### 2022-02-25 #4148 @slwu89
@fboehm I have a question, in the documentation section of the review checking it indicates:

> A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?

Does this mean that the README file needs to have a section titled "Statement of need" or is it all right if this information is available but not labeled as such?
