### 2022-02-25 arcaldwell49/SimplyAgree#9 @arcaldwell49
Just pushed an update in the branch I am working on where the website is linked to in the README. Just as a note, the URL is already in URL field of the DESCRIPTION file.

Undergoing checks now and should be merged with the main branch today.
