### 2022-02-27 1720754
Checking MacBook
Checked on local MacBook. Running fine. Appears to be issue with GitHub jobs.
