### 2022-02-27 arcaldwell49/SimplyAgree#10 @arcaldwell49
Branch added with code of conduct and contribution guide.
