### 2022-02-27 4221044
Citations
Added citations for R packages with similiar functions. Need to add reference of them to the paper. Must emphasize the GUI part of the package.
