+++ paper/paper.bib
@@ -149,3 +149,45 @@
   version = {1.6.23},
   date = {2021-05-11},
 }
+
+@Manual{MethComp,
+    title = {MethComp: Analysis of Agreement in Method Comparison Studies},
+    author = {Bendix Carstensen and Lyle Gurrin and Claus Thorn Ekstrøm and Michal Figurski},
+    year = {2020},
+    note = {R package version 1.30.0},
+    url = {https://CRAN.R-project.org/package=MethComp},
+  }
+
+  @Manual{blandr,
+    title = {blandr: a Bland-Altman Method Comparison package for R},
+    author = {Deepankar Datta},
+    doi = {10.5281/zenodo.824514},
+    year = {2017},
+    url = {https://github.com/deepankardatta/blandr},
+  }
+
+@Manual{BlandAltmanLeh,
+    title = {BlandAltmanLeh: Plots (Slightly Extended) Bland-Altman Plots},
+    author = {Bernhard Lehnert},
+    year = {2015},
+    note = {R package version 0.3.1},
+    url = {https://CRAN.R-project.org/package=BlandAltmanLeh},
+  }
+
+ @Manual{psych,
+    title = {psych: Procedures for Psychological, Psychometric, and Personality Research},
+    author = {William Revelle},
+    organization = { Northwestern University},
+    address = { Evanston, Illinois},
+    year = {2020},
+    note = {R package version 2.0.12},
+    url = {https://CRAN.R-project.org/package=psych},
+  }
+
+  @Article{ICC,
+    title = {Guidelines for Estimating Repeatability},
+    author = {Matthew E. Wolak and Daphne J. Fairbairn and Yale R. Paulsen},
+    year = {2012},
+    journal = {Methods in Ecology and Evolution 3(1):129-137},
+  }
+
