### 2022-02-28 65a9f14
Added footnote
Footnote for current R capabilities
