+++ paper/paper.bib
@@ -30,7 +30,7 @@
   volume = {327},
   number = {8476},
   pages = {307--310},
-  author = {J. Martin Bland and DouglasG. Altman},
+  author = {J Martin Bland and DouglasG. Altman},
   title = {Statistical methods for assessing agreement between two methods of clinical measurement},
   journal = {The Lancet}
 }
@@ -85,7 +85,7 @@
   volume = {26},
   number = {16},
   pages = {3095--3113},
-  author = {Tonya S. King and Vernon M. Chinchilli and Josep L. Carrasco},
+  author = {Tonya S King and Vernon M Chinchilli and Josep L Carrasco},
   title = {A repeated measures concordance correlation coefficient},
   journal = {Statistics in Medicine}
 }
@@ -99,7 +99,7 @@
   volume = {19},
   number = {1},
   pages = {90--105},
-  author = {Josep L. Carrasco and Tonya S. King and Vernon M. Chinchilli},
+  author = {Josep L Carrasco and Tonya S King and Vernon M Chinchilli},
   title = {The Concordance Correlation Coefficient for Repeated Measures Estimated by Variance Components},
   journal = {Journal of Biopharmaceutical Statistics}
 }
@@ -113,7 +113,7 @@
   volume = {109},
   number = {3},
   pages = {293--304},
-  author = {Josep L. Carrasco and Brenda R. Phillips and Josep Puig-Martinez and Tonya S. King and Vernon M. Chinchilli},
+  author = {Josep L Carrasco and Brenda R Phillips and Josep Puig-Martinez and Tonya S King and Vernon M Chinchilli},
   title = {Estimation of the concordance correlation coefficient for repeated measures using {SAS} and R},
   journal = {Computer Methods and Programs in Biomedicine}
 }
@@ -186,7 +186,7 @@
 
   @Article{ICC,
     title = {Guidelines for Estimating Repeatability},
-    author = {Matthew E. Wolak and Daphne J. Fairbairn and Yale R. Paulsen},
+    author = {Matthew E Wolak and Daphne J Fairbairn and Yale R Paulsen},
     year = {2012},
     journal = {Methods in Ecology and Evolution 3(1):129-137},
   }
