### 2022-02-28 f01d866
Fix errors again
NA values problem with agree_reps resolved  Dependencies also reduced.
