### 2022-02-28 e9a4e66
Update to paper
Working on reviewer comments and making the paper/package's intention clearer.
