### 2022-02-28 6524129
Start of jamovi update
Updating jamovi, reduce depends and improve UI
