+++ build/R4.0.5-win64/SimplyAgree/NEWS.md
@@ -1,0 +1,11 @@
+# NEWS
+
+## v0.0.3
+- Fixed error in `agree_nest` and `agree_reps`; now properly handles missing values
+- Remove dependencies on sjstats and cccrm packages
+
+## v0.0.2
+- Fixes typos in jamovi/jmv functions
+- Adds more descriptive errors to jamovi output
+- Remove dontrun from examples in documentation
+- Add more details and references to the package's functions
