+++ build/R4.0.5-win64/SimplyAgree/NAMESPACE
@@ -1,0 +1,60 @@
+# Generated by roxygen2: do not edit by hand
+
+S3method(plot,loa_mixed_bs)
+S3method(plot,powerCurve)
+S3method(plot,simple_agree)
+S3method(plot,simple_reli)
+S3method(print,loa_mixed_bs)
+S3method(print,simple_agree)
+S3method(print,simple_reli)
+export(agree_nest)
+export(agree_reps)
+export(agree_test)
+export(blandPowerCurve)
+export(find_n)
+export(jmvagree)
+export(jmvagreemulti)
+export(jmvreli)
+export(loa_mixed)
+export(reli_stats)
+import(boot)
+import(dplyr)
+import(emmeans)
+import(ggplot2)
+import(lme4)
+importFrom(dplyr,mutate)
+importFrom(dplyr,rename)
+importFrom(dplyr,select)
+importFrom(magrittr,"%>%")
+importFrom(stats,anova)
+importFrom(stats,aov)
+importFrom(stats,as.formula)
+importFrom(stats,complete.cases)
+importFrom(stats,cor)
+importFrom(stats,dchisq)
+importFrom(stats,lm)
+importFrom(stats,na.omit)
+importFrom(stats,pf)
+importFrom(stats,pnorm)
+importFrom(stats,pt)
+importFrom(stats,qchisq)
+importFrom(stats,qf)
+importFrom(stats,qnorm)
+importFrom(stats,qt)
+importFrom(stats,residuals)
+importFrom(stats,sd)
+importFrom(stats,var)
+importFrom(stringr,str_split)
+importFrom(tidyr,drop_na)
+importFrom(tidyr,pivot_longer)
+importFrom(tidyselect,all_of)
+importFrom(utils,stack)
+
+export(jmvagreeClass)
+export(jmvagreeOptions)
+
+export(jmvagreemultiClass)
+export(jmvagreemultiOptions)
+
+export(jmvreliClass)
+export(jmvreliOptions)
