### 2022-03-01 arcaldwell49/SimplyAgree#16 @arcaldwell49
Reduce Dependencies
Suggested reduction in dependencies https://github.com/openjournals/joss-reviews/issues/4148#issuecomment-1055113260

I completely agree and I am already working to remove the sjstats and cccrm dependencies due to some newly reported issues regarding these packages.

Putting in pull request later today.
