### 2022-03-01 #4148 @arcaldwell49
@itismeghasyam  `cccrm` and `sjstats` have been removed as required dependencies for v0.0.3 now on the main branch.

`cccrm` is only suggested for the purposes of the vignettes.
