### 2022-03-02 bf73d49
Start of agree_coef functions
Working on weights and coefficients.
