### 2022-03-02 afbabc6
Update coef_funcs
