+++ R/coef_funcs.R
@@ -22,7 +22,7 @@ quadratic_weights = function (categ)
   }
 
 
-pa_coef = function(ratings.mat){
+pa_coef = function(ratings.mat, conf.level){
 
   if (is.character(ratings.mat)){
     ratings.mat <- trim(toupper(ratings.mat))
@@ -55,15 +55,42 @@ pa_coef = function(ratings.mat){
   }
   agree.mat.w <- t(weights.mat%*%t(agree.mat))
 
-  # calculating percent agreement coefficient
+  # calculating percent agreement coeficient
 
   ri.vec <- agree.mat%*%rep(1,q)
   sum.q <- (agree.mat*(agree.mat.w-1))%*%rep(1,q)
   n2more <- sum(ri.vec>=2)
   pa <- sum(sum.q[ri.vec>=2]/((ri.vec*(ri.vec-1))[ri.vec>=2]))/n2more
   pe <- 0
-  coeff.val <- pa
+  coef.val <- pa
 
+  if (q>=2){
+    # calculating variance, stderr & p-value of percent agreement
+
+    den.ivec <- ri.vec*(ri.vec-1)
+    den.ivec <- den.ivec - (den.ivec==0) # this operation replaces each 0 value with -1 to make the next ratio calculation always possible.
+    pa.ivec <- (n/n2more)*(sum.q/den.ivec)
+    var.pa<-NA;stderr<-NA;stderr.est<-NA;p.value<-NA;lcb<-NA;ucb<-NA
+    if (n>=2){
+      var.pa <- ((1-f)/(n*(n-1))) * sum((pa.ivec - pa)^2)
+      stderr <- sqrt(var.pa)# pa's standard error
+      stderr.est <- round(stderr,5)
+      p.value <- 1-pt(pa/stderr,n-1)
+      lcb <- pa - stderr*qt(1-(1-conf.level)/2,n-1) # lower confidence bound
+      ucb <- min(1,pa + stderr*qt(1-(1-conf.level)/2,n-1)) # upper confidence bound
+    }
+    #conf.int <- paste0("(",round(lcb,3),",",round(ucb,3),")")
+    coef.se <- stderr.est
+  }
+  cnames <- colnames(ratings)
+  colnames(ratings) <- sapply(1:r, function(x) paste0("dumcol", x))
+  obs.count <- dplyr::summarise(as.data.frame(1-is.na(ratings)),across(1:r,sum))
+  tot.obs <- sum((1-is.na(ratings)))
+  colnames(obs.count) <- cnames
+  coef.name <- "Percent Agreement"
+  df.out <- data.frame(coef.name,pa,pe,coef.val,coef.se,
+                       coef.lower = lcb, coef.upper = ucb,p.value)
+  return(df.out)
 
 
 }
