### 2022-03-02 227348d
Update coef_funcs.R
