+++ R/coef_funcs.R
@@ -22,7 +22,9 @@ quadratic_weights = function (categ)
   }
 
 
-pa_coef = function(ratings.mat, conf.level){
+pa_coef = function(ratings.mat,
+                   conf.level,
+                   weighted =FALSE){
 
   if (is.character(ratings.mat)){
     ratings.mat <- trim(toupper(ratings.mat))
@@ -30,7 +32,6 @@ pa_coef = function(ratings.mat, conf.level){
   }
   n <- nrow(ratings.mat) # number of subjects
   r <- ncol(ratings.mat) # number of raters
-  f <- 0 # finite population correction
 
   # creating a vector containing all categories used by the raters
 
@@ -41,11 +42,11 @@ pa_coef = function(ratings.mat, conf.level){
 
   # creating the weights matrix
 
-    if (weights==TRUE) {
-      weights.mat<-quadratic_weights(categ)
-    } else {
-      weights.mat<-diag(length(categ))
-    }
+  if (weighted == TRUE) {
+    weights.mat <- quadratic_weights(categ)
+  } else {
+    weights.mat <- diag(length(categ))
+  }
   # creating the nxq agreement matrix representing the distribution of raters by subjects and category
 
   agree.mat <- matrix(0,nrow=n,ncol=q)
@@ -62,35 +63,74 @@ pa_coef = function(ratings.mat, conf.level){
   n2more <- sum(ri.vec>=2)
   pa <- sum(sum.q[ri.vec>=2]/((ri.vec*(ri.vec-1))[ri.vec>=2]))/n2more
   pe <- 0
-  coef.val <- pa
+  coef_val <- pa
+
+  den.ivec <- ri.vec*(ri.vec-1)
+  den.ivec <- den.ivec - (den.ivec==0) # this operation replaces each 0 value with -1 to make the next ratio calculation always possible.
+  pa.ivec <- (n/n2more)*(sum.q/den.ivec)
 
   if (q>=2){
     # calculating variance, stderr & p-value of percent agreement
 
-    den.ivec <- ri.vec*(ri.vec-1)
-    den.ivec <- den.ivec - (den.ivec==0) # this operation replaces each 0 value with -1 to make the next ratio calculation always possible.
-    pa.ivec <- (n/n2more)*(sum.q/den.ivec)
-    var.pa<-NA;stderr<-NA;stderr.est<-NA;p.value<-NA;lcb<-NA;ucb<-NA
+
+
     if (n>=2){
-      var.pa <- ((1-f)/(n*(n-1))) * sum((pa.ivec - pa)^2)
+      var.pa <- ((1)/(n*(n-1))) * sum((pa.ivec - pa)^2)
       stderr <- sqrt(var.pa)# pa's standard error
       stderr.est <- round(stderr,5)
       p.value <- 1-pt(pa/stderr,n-1)
       lcb <- pa - stderr*qt(1-(1-conf.level)/2,n-1) # lower confidence bound
       ucb <- min(1,pa + stderr*qt(1-(1-conf.level)/2,n-1)) # upper confidence bound
     }
-    #conf.int <- paste0("(",round(lcb,3),",",round(ucb,3),")")
-    coef.se <- stderr.est
+
+    coef_se <- stderr.est
   }
-  cnames <- colnames(ratings)
-  colnames(ratings) <- sapply(1:r, function(x) paste0("dumcol", x))
-  obs.count <- dplyr::summarise(as.data.frame(1-is.na(ratings)),across(1:r,sum))
[…]
