### 2022-03-04 0da777c
Update coef_funcs.R
