### 2022-03-06 #4148 @fboehm
@arcaldwell49 - the pdf paper looks really good! please make the minor corrections below.

A few minor fixes are needed in the pdf. I think that these can be accomplished by updating the bib file as needed.

line 110: R needs to be capitalized. You can accomplish this by putting braces around capital R in your bib file, like {R}
line 115: capitalize R.
line 117: Do you need to capitalize Bland-Altman? I'm guessing you do.
line 122: Capitalize Bland-Altman
line 139: Capitalize BlandAltman

Once these fixes are made, we can proceed towards publication.
