### 2022-03-07 #4148 @arcaldwell49
@fboehm good catches on my dumb typos!

All requested typos fixed.
