### 2022-03-07 #4148 @arcaldwell49
Oh wow, good catch! It seems JSCR has resolved the doi issue. Pushing new version in a minute...
