### 2022-03-08 #4148 @arcaldwell49
Hi @fboehm, I will make the tagged release ASAP. As far as the archive goes, are you wanting something like a Zenodo archive of the paper.md file?
