### 2022-03-08 #4148 @arcaldwell49
@fboehm Release is up. https://github.com/arcaldwell49/SimplyAgree/releases/tag/0.0.3

 I have a draft in Zenodo of v0.0.3 that I will publish with your confirmation that this is how you would like me to proceed.
