### 2022-03-09 #4148 @fboehm
Thanks, @arcaldwell49 . The archive should contain the entire repository, not just the paper.md file. I think that what you've shared is what's needed. Please correct me if I'm mistaken.
