### 2022-03-12 #4148 @Kevin-Mattheus-Moerman
@openjournals/dev @xuanxu @arfon I think we have an issue with multiple affiliations for one author. Here the paper features the yaml `affiliation: "1, 2"` However, the affiliation appears as superscript `12` rather than `1, 2`.

![Screenshot from 2022-03-12 08-38-14](https://user-images.githubusercontent.com/8392709/158010764-81f0f8c5-6eb1-4fd5-bda7-d07f3347390e.png)
