+++ cran-comments.md
@@ -1,4 +1,3 @@
 # Comments
 
-- dontrun removed from documentation
-- No reference is included in the description at this time. The relevant reference for each function is the documentation for each function.
+- Submission is to fix small error in agree_nest function
