+++ cran-comments.md
@@ -1,3 +1,3 @@
 # Comments
 
-- Submission is to fix small error in agree_nest function
+- README URLs fixed.
