### 2022-03-14 854a5a8
Added check_ba method
