+++ NAMESPACE
@@ -1,5 +1,6 @@
 # Generated by roxygen2: do not edit by hand
 
+S3method(check_ba,simple_agree)
 S3method(plot,loa_mixed_bs)
 S3method(plot,powerCurve)
 S3method(plot,simple_agree)
@@ -11,6 +12,7 @@ export(agree_nest)
 export(agree_reps)
 export(agree_test)
 export(blandPowerCurve)
+export(check_ba)
 export(find_n)
 export(jmvagree)
 export(jmvagreemulti)
@@ -33,8 +35,11 @@ importFrom(stats,as.formula)
 importFrom(stats,complete.cases)
 importFrom(stats,cor)
 importFrom(stats,dchisq)
+importFrom(stats,df.residual)
+importFrom(stats,ks.test)
 importFrom(stats,lm)
 importFrom(stats,na.omit)
+importFrom(stats,pchisq)
 importFrom(stats,pf)
 importFrom(stats,pnorm)
 importFrom(stats,pt)
@@ -43,7 +48,11 @@ importFrom(stats,qf)
 importFrom(stats,qnorm)
 importFrom(stats,qt)
 importFrom(stats,residuals)
+importFrom(stats,rstandard)
+importFrom(stats,rstudent)
 importFrom(stats,sd)
+importFrom(stats,shapiro.test)
+importFrom(stats,sigma)
 importFrom(stats,var)
 importFrom(stringr,str_split)
 importFrom(tidyr,drop_na)
