### 2022-03-15 66f81f0
Adding smooths to plots
