+++ R/agree_test.R
@@ -7,7 +7,8 @@
 #' @param delta The threshold below which methods agree/can be considered equivalent, can be in any units. Often referred to as the "Equivalence Bound for Agreement" or "Maximal Allowable Difference".
 #' @param x_lab Label for x values (first measurement)
 #' @param y_lab Label for y values (second measurement)
-#'
+#' @param smooth_method Smoothing method (function) to use, accepts either NULL or a character vector, e.g. "lm", "glm", "gam", "loess" or a function. Default is NULL, which will not include a trend line.
+#' @param smooth_se Display confidence interval around smooth?
 #' @return Returns single list with the results of the agreement analysis.
 #'
 #' \describe{
@@ -46,7 +47,9 @@ agree_test <- function(x,
                        conf.level = .95,
                        agree.level = .95,
                        x_lab = "x",
-                       y_lab = "y") {
+                       y_lab = "y",
+                       smooth_method = NULL,
+                       smooth_se = TRUE) {
   est <- lower.ci <- upper.ci <- NULL
   if (agree.level >= 1 || agree.level <= 0) {
 
@@ -197,6 +200,21 @@ agree_test <- function(x,
       scale_y_continuous(sec.axis = dup_axis(
         breaks = c(delta, -1*delta),
         name = "Maximal Allowable Difference"))
+  }
+  if (!is.null(smooth_method)){
+    if (!(smooth_method %in% c("loess", "lm", "gam"))){
+      stop("Only lm, loess, and gam are supported as smooth_method at this time.")
+    }
+    bland_alt.plot = bland_alt.plot +
+      stat_smooth(
+        method = smooth_method,
+        se = smooth_se,
+        level = conf.level,
+        alpha = 0.2,
+        formula = y ~ x,
+        size = 0.8,
+        colour = "#3aaf85"
+      )
   }
 
 
