### 2022-03-15 a7a29b5
Add deming line
Also adding options for gam and loess for agree_reps next
