+++ NAMESPACE
@@ -42,6 +42,7 @@ importFrom(stats,na.omit)
 importFrom(stats,pchisq)
 importFrom(stats,pf)
 importFrom(stats,pnorm)
+importFrom(stats,prcomp)
 importFrom(stats,pt)
 importFrom(stats,qchisq)
 importFrom(stats,qf)
