### 2022-03-16 f767b6c
Added methods for reli_stats
