+++ R/methods.simple_reli.R
@@ -48,3 +48,13 @@ plot.simple_reli <- function(x,  ...){
   return(x$plot.reliability)
 
 }
+
+
+#' @rdname simple_reli-methods
+#' @method check simple_reli
+#' @importFrom stats residuals lm na.omit pchisq shapiro.test ks.test rstudent df.residual anova rstandard sigma resid
+#' @export
+
+check.simple_reli <- function(x) {
+  check_reli(x)
+}
