+++ tests/testthat/test_agree_test.R
@@ -14,7 +14,7 @@ testthat::test_that("Simple Use Run Through", {
   agree1 = agree_test(x = reps$x,
                       y = reps$y,
                       delta = 2.5,
-                      smooth_method = "lm",
+                      smooth_method = "gam",
                       smooth_se = FALSE)
 
   jmvagree1 = jmvagree(data = reps,
@@ -29,8 +29,8 @@ testthat::test_that("Simple Use Run Through", {
 
   p = plot(agree1,type=1)
   p = plot(agree1,type=2)
-  c1 = check_ba(agree1)$p_norm
-  c2 = check_ba(agree1)$p_het
+  c1 = check(agree1)$p_norm
+  c2 = check(agree1)$p_het
 
   testthat::expect_equal(agree1$ccc.xy$est.ccc,
                          .479,
