### 2022-03-16 7f16354
Add deming to other functions
