+++ NAMESPACE
@@ -37,6 +37,7 @@ importFrom(stats,complete.cases)
 importFrom(stats,cor)
 importFrom(stats,dchisq)
 importFrom(stats,df.residual)
+importFrom(stats,fitted)
 importFrom(stats,ks.test)
 importFrom(stats,lm)
 importFrom(stats,na.omit)
