### 2022-03-17 f45ab7a
Simple plot ba function working
Must add modifications for replicate data.
