+++ R/simple_plot_functions.R
@@ -1,24 +1,25 @@
 
 simple_ident_plot = function(x) {
-  if(x$class == "replicates"){
-    form1 = as.formula(paste0(x$call$x, "~",
-                              x$call$y, " + ",
-                              x$call$id))
-    call2 = structure(list(formula = form1,
-                 data = x$call$data), class = "call")
+  if(x$class != "simple"){
+    df = model.frame(x$call)
+    colnames(df) = c("x","y","id")
+    if(x$class == "replicates"){
     df = df %>%
       group_by(id) %>%
       summarize(mxi = sum(!is.na(x)),
                 myi = sum(!is.na(y)),
-                x_bar = mean(x, na.rm=TRUE),
+                x = mean(x, na.rm=TRUE),
                 x_var = var(x, na.rm=TRUE),
-                y_bar = mean(y, na.rm=TRUE),
+                y = mean(y, na.rm=TRUE),
                 y_var = var(y, na.rm=TRUE),
                 .groups = "drop")
+    }
 
+  } else{
+    df = model.frame(x$call)
   }
   #df_loa = x$loa
-  df = model.frame(x$call)
+
   scalemin = min(c(min(df$x, na.rm = TRUE),min(df$y, na.rm = TRUE)))
   scalemax = max(c(max(df$x, na.rm = TRUE),max(df$y, na.rm = TRUE)))
   x_lab = x$labs$x_lab
@@ -56,4 +57,92 @@ simple_ident_plot = function(x) {
 }
 
 
+simple_ba_plot = function(x) {
+  if(x$class != "simple"){
+    df = model.frame(x$call)
+    colnames(df) = c("x","y","id")
 
+  } else{
+    df = model.frame(x$call)
+  }
+  df_loa = x$loa
+
+  scalemin = min(c(min(df$x, na.rm = TRUE),min(df$y, na.rm = TRUE)))
+  scalemax = max(c(max(df$x, na.rm = TRUE),max(df$y, na.rm = TRUE)))
+  pd2 = position_dodge2(.03*(scalemax-scalemin))
+  x_lab = x$labs$x_lab
+  y_lab = x$labs$y_lab
+  df_loa2 = df_loa
+  df_loa2$x = scalemin
+  df_loa2$text = factor(c("Bias", "Lower LoA", "Upper LoA"),
+                       levels = c("Upper LoA", "Bias", "Lower LoA"))
+  df$mean = (df$x + df$y)/2
+  df$delta = df$x - df$y
+  conf.level = x$conf.level
+  agree.level = x$agree.level
+  delta = x$delta
+  smooth_method = x$smooths$smooth_method
+  smooth_se = x$smooths$smooth_se
+  ggplot(df,
+         aes(x = mean, y = delta)) +
+    geom_point(na.rm = TRUE) +
+    geom_pointrange(data = df_loa2,
+                    aes(
+                      x = x,
+                      y = estimate,
+                      ymin = lower.ci,
+                      ymax = upper.ci,
+                      color = text),
+                    #width = .03*(scalemax-scalemin),
+                    position = pd2,
+                    inherit.aes = FALSE)+
+    labs(x = paste0("Average of ", x_lab ," & ", y_lab),
+         y = paste0("Difference between Methods ",x_lab ," & ", y_lab),
+         caption = paste0("Agreement = ", agree.level * 100,"% \n",
+                          "Confidence Level = ", conf.level * 100, "%"),
+         color = "") +
+    scale_color_viridis_d(option = "C", end = .8) +
+    theme_bw() +
+    theme(legend.position = "left")
+  if (!is.null(x$delta)){
+    delta = x$delta
+    df_delta = data.frame(y1 = c(delta, -1*delta))
+    bland_alt.plot = bland_alt.plot +
[…]
