+++ R/other_functions.R
@@ -400,156 +400,6 @@ plot_bias = function(x,
 
 # Check agree_test -----
 
-check_simple = function(x){
-
-  dat = x$bland_alt.plot$data
-
-  ## Heteroskedasticity -------
-  mod_check = lm(data = dat,
-                 delta ~ 1)
-  stan_res = residuals(mod_check, type = "pearson")
-  df_het = df.residual(mod_check)
-  sum_het_res = sum(!is.na(stan_res))
-  sigma_het = sigma(mod_check)
-
-  s_sq = df_het * sigma_het^2 / sum_het_res
-
-  u_het = stan_res^2 / s_sq
-
-  mod <- lm(u_het ~ dat$mean)
-
-  SS <- anova(mod)$"Sum Sq"
-  RegSS <- sum(SS) - SS[length(SS)]
-  Chisq <- RegSS / 2
-  ### Breusch-Pagan Test
-  p_val_het <- pchisq(Chisq, df = 1, lower.tail = FALSE)
-
-  rstan_het =  rstandard(mod_check)
-  dat_het <- data.frame(
-    x = dat$mean,
-    y = sqrt(abs(rstan_het))
-  )
-  p_het = plot_het(dat_het) +
-    labs(caption = paste0("Heteroskedasticity", " \n",
-                                   "Breusch-Pagan Test: p = ",
-                                   signif(p_val_het,4)))
-
-
-  ## Normality ------------
-
-  mod_res = residuals(mod_check)
-  if(length(mod_res) < 5000){
-    norm_test = shapiro.test(mod_res)
-    norm_text = "Shapiro-Wilk Test"
-  } else {
-    norm_test = ks.test(mod_res, y = "pnorm",
-                               alternative = "two.sided")
-    norm_text = "Kolmogorov-Smirnov Test"
-  }
-
-  rstan_norm = sort(rstudent(mod_check), na.last = NA)
-  dat_norm <- na.omit(data.frame(y = rstan_norm))
-  p_norm = plot_qq(
-    x = dat_norm
-  ) +
-    labs(caption = paste0("Normality", " \n",
-                                   norm_text, ": p = ",
-                                   signif(p_val_het,4)))
-
-  # Proportional Bias ----
-  mod2 = lm(delta ~ mean,
-            data = dat)
-  aov2 = as.data.frame(anova(mod_check, mod2))
-  colnames(aov2) = c("df1","RSS","df2","SS","f","p")
-  lin_pval = aov2$p[2]
-  dat2 = data.frame(resid = residuals(mod_check),
-                    mean = na.omit(dat$mean))
-  p_bias = plot_bias(dat2) +
-    labs(caption = paste0("Proportional Bias", " \n",
-                          "Test for Linear Bias", ": p = ",
-                          signif(lin_pval,4)))
-
-  return(list(p_norm = p_norm,
-              p_het = p_het,
-              p_bias = p_bias))
-}
-
-
-check_multi = function(x){
-
-
-  dat = x$bland_alt.plot$data
-
-  ## Heteroskedasticity -------
-  mod_check = lme4::lmer(data = dat,
-                         d ~ 1 + (1 | id))
-  stan_res = residuals(mod_check, type = "pearson")
-  df_het = df.residual(mod_check)
-  sum_het_res = sum(!is.na(stan_res))
-  sigma_het = sigma(mod_check)
-
-  s_sq = df_het * sigma_het^2 / sum_het_res
-
-  u_het = stan_res^2 / s_sq
-
-  mod <- lm(u_het ~ na.omit(dat$avg_both))
-
-  SS <- anova(mod)$"Sum Sq"
-  RegSS <- sum(SS) - SS[length(SS)]
-  Chisq <- RegSS / 2
-  ### Breusch-Pagan Test
[…]
