+++ R/agree_test.R
@@ -5,10 +5,6 @@
 #' @param conf.level the confidence level required. Default is 95\%.
 #' @param agree.level the agreement level required. Default is 95\%. The proportion of data that should lie between the thresholds, for 95\% limits of agreement this should be 0.95.
 #' @param delta The threshold below which methods agree/can be considered equivalent, can be in any units. Often referred to as the "Equivalence Bound for Agreement" or "Maximal Allowable Difference".
-#' @param x_lab Label for x values (first measurement)
-#' @param y_lab Label for y values (second measurement)
-#' @param smooth_method Smoothing method (function) to use, accepts either NULL or a character vector, e.g. "lm", "glm", "gam", "loess" or a function. Default is NULL, which will not include a trend line.
-#' @param smooth_se Display confidence interval around smooth?
 #' @return Returns single list with the results of the agreement analysis.
 #'
 #' \describe{
@@ -45,11 +41,7 @@ agree_test <- function(x,
                        y,
                        delta,
                        conf.level = .95,
-                       agree.level = .95,
-                       x_lab = "x",
-                       y_lab = "y",
-                       smooth_method = NULL,
-                       smooth_se = TRUE) {
+                       agree.level = .95) {
   est <- lower.ci <- upper.ci <- NULL
   if (agree.level >= 1 || agree.level <= 0) {
 
@@ -244,10 +236,6 @@ agree_test <- function(x,
                  h0_test = rej_text,
                  call = lm_mod,
                  delta = delta,
-                 smooths = list(smooth_method = smooth_method,
-                                smooth_se = smooth_se),
-                 labs = list(x_lab = x_lab,
-                             y_lab = y_lab),
                  class = "simple"),
             class = "simple_agree")
 
