+++ R/methods.simple_agree.R
@@ -108,15 +108,33 @@ print.simple_agree <- function(x,...){
 
 #' @rdname simple_agree-methods
 #' @method plot simple_agree
+#' @param x_name Name/label for x values (first measurement)
+#' @param y_name Name/label for y values (second measurement)
+#' @param smooth_method Smoothing method (function) to use, accepts either NULL or a character vector, e.g. "lm", "glm", "gam", "loess" or a function. Default is NULL, which will not include a trend line.
+#' @param smooth_se Display confidence interval around smooth?
 #' @import ggplot2
 #' @export
 
-plot.simple_agree <- function(x, type = 1, ...){
+plot.simple_agree <- function(x, type = 1,
+                              x_name = "x",
+                              y_name = "y",
+                              smooth_method = NULL,
+                              smooth_se = TRUE,
+                              ...){
 
   if(type == 1){
-    return(x$bland_alt.plot)
+    #return(x$bland_alt.plot)
+    simple_ba_plot(x,
+                   x_name,
+                   y_name,
+                   smooth_method,
+                   smooth_se)
   } else if (type == 2){
-    simple_ident_plot(x)
+    simple_ident_plot(x,
+                      x_name,
+                      y_name,
+                      smooth_method,
+                      smooth_se)
   } else{
    stop("please select type = 1 for a Bland Altman plot or type = 2 for an identity plot")
   }
@@ -137,12 +155,100 @@ check <- function(x) {
 #' @export
 
 check.simple_agree <- function(x) {
-  if (x$class == "simple") {
-    check_simple(x)
-  } else if(x$class == "nested" || x$class == "replicates") {
-    if(x$class == "nested"){warning("Warning: assumptions tests for agree_nest are only approximate. Proceed with caution.")}
-    check_multi(x)
-  } else {
-    stop("Only simple, nested, or replicates supported at this time.")
+
+  if(x$class == "nested"){warning("Warning: assumptions tests for agree_nest are only approximate. Proceed with caution.")}
+  if(x$class != "simple"){
+    df = model.frame(x$call)
+    colnames(df) = c("x","y","id")
+  } else{
+    df = model.frame(x$call)
   }
+  df$mean = (df$x + df$y)/2
+  df$delta = df$x - df$y
+
+  dat = df
+  ## Heteroskedasticity -------
+  mod_check = if (x$class != "simple") {
+    lme4::lmer(data = dat,
+               delta ~ 1 + (1 | id))
+  } else {
+    lm(data = dat,
+       delta ~ 1)
+  }
+
+  stan_res = residuals(mod_check, type = "pearson")
+  df_het = df.residual(mod_check)
+  sum_het_res = sum(!is.na(stan_res))
+  sigma_het = sigma(mod_check)
+
+  s_sq = df_het * sigma_het^2 / sum_het_res
+
+  u_het = stan_res^2 / s_sq
+
+  mod <- lm(u_het ~ na.omit(dat$mean))
+
+  SS <- anova(mod)$"Sum Sq"
+  RegSS <- sum(SS) - SS[length(SS)]
+  Chisq <- RegSS / 2
+  ### Breusch-Pagan Test
+  p_val_het <- pchisq(Chisq, df = 1, lower.tail = FALSE)
+
+  rstan_het =  residuals(mod_check, scaled = TRUE)
+  dat_het <- data.frame(
+    x = na.omit(dat$mean),
+    y = na.omit(sqrt(abs(rstan_het)))
+  )
+  p_het = plot_het(dat_het) +
+    labs(caption = paste0("Heteroskedasticity", " \n",
[…]
