### 2022-03-18 fc8898e
reli_stats updated
plot function now works off call
