### 2022-03-18 ff5cd97
Add deming fucntions
