### 2022-03-19 5f21426
agree_np still working on it
Make it simpe-agree object.

Will need to adjust checks. The LoA with prop_bias = TRUE can be from the ends and median
