+++ build/R4.1.2-win64/SimplyAgree/DESCRIPTION
@@ -1,0 +1,21 @@
+Package: SimplyAgree
+Type: Package
+Title: Flexible and Robust Agreement and Reliability Analyses
+Version: 0.1.0
+Authors@R: person("Aaron", "Caldwell", email = "arcaldwell49@gmail.com",
+                  role = c("aut", "cre"))
+Maintainer: Aaron Caldwell <arcaldwell49@gmail.com>
+Description: Reliability and agreement analyses often have limited software support. Therefore, this package was created to make agreement and reliability analyses easier for the average researcher. The functions within this package include simple tests of agreement, agreement analysis for nested and replicate data, and provide robust analyses of reliability. In addition, this package contains a set of functions to help when planning studies looking to assess measurement agreement. For robust analyses of agreement, limits of agreement through a bootstrap method can also be calculated.
+URL: https://aaroncaldwell.us/SimplyAgree/
+License: GPL (>= 3)
+Encoding: UTF-8
+LazyData: true
+RoxygenNote: 7.1.2
+Imports: ggplot2, emmeans, lme4, boot, stats, dplyr, magrittr,
+        tidyselect, tidyr, stringr, jmvcore, quantreg
+Suggests: knitr, rmarkdown, testthat, tidyverse, readr, ggpubr, covr,
+        cccrm, mgcv, ggeffects
+VignetteBuilder: knitr
+Depends: R (>= 3.5)
+Author: Aaron Caldwell [aut, cre]
+Built: R 4.1.2; ; 2022-03-19 16:14:55 UTC; windows
