### 2022-03-20 40d94b3
working on proportional bias now
