+++ R/agree_np.R
@@ -12,11 +12,11 @@
 #'
 #' \describe{
 #'   \item{\code{"loa"}}{a data frame of the limits of agreement.}
+#'   \item{\code{"agee"}}{a data frame of the binomial proportion of results in agreement..}
 #'   \item{\code{"h0_test"}}{Decision from hypothesis test.}
 #'   \item{\code{"call"}}{the matched call}
 #'
 #' }
-
 #' @examples
 #' # TBA
 #' @section References:
@@ -44,6 +44,7 @@ agree_np <- function(x,
   confq = qnorm(1 - (1 - conf.level) / 2)
   alpha.l = 1 - (1 - conf.level) / 2
   alpha.u = (1 - conf.level) / 2
+  stopifnot(!is.null(delta) && is.numeric(delta) && delta > 0)
 
   if(!is.null(id)){
     df = data %>%
@@ -51,7 +52,8 @@ agree_np <- function(x,
       rename(id = all_of(id),
              x = all_of(x),
              y = all_of(y)) %>%
-      select(id,x,y)
+      select(id,x,y) %>%
+      drop_na()
   } else {
     df = data %>%
       select(all_of(x),all_of(y)) %>%
@@ -61,50 +63,162 @@ agree_np <- function(x,
     df$id = as.factor(1)
     df = df %>%
       drop_na()
+
   }
 
   df$mean = (df$x+df$y)/2
   df$delta = (df$x-df$y)
+  dbound = delta
+  if(!is.null(id)){
+    df_glm1 = df
+    df_glm1 = df_glm1 %>%
+      mutate(success = ifelse(abs(delta) <= dbound,
+                              1,0),
+             failure = ifelse(abs(delta) > dbound,
+                              1,0))
 
-  quan_mod2 = rq(formula =  delta ~ x,
-                data = df,
-                tau = c(agree.u,.5,agree.l))
+    df_glm = df_glm1 %>%
+      group_by(id) %>%
+      summarize(x = mean(x),
+                y = mean(y),
+                success = sum(success),
+                failure = sum(failure))
 
-
-  quan_mod1 = suppressWarnings({rq(formula =  delta ~ 1,
-                 data = df,
-                 tau = c(agree.u,.5,agree.l))})
-  df_coef = data.frame(est = c(),
-                       se = c())
-  quan_coef = summary(quan_mod1, se = "boot")
-  for(i in 1:length(quan_coef)){
-    temp = (as.data.frame(quan_coef[[i]]$coefficients))
-    df_coef = rbind(df_coef,temp[,1:2])
+  }else{
+    df_glm = df
+    df_glm = df_glm %>%
+      mutate(success = ifelse(abs(delta) <= dbound,
+             1,0),
+             failure = ifelse(abs(delta) > dbound,
+                                        1,0))
   }
-  df_coef = data.frame(est = df_coef$Value,
-                       se = df_coef$`Std. Error`)
-  rownames(df_coef) = c("Lower LoA", "Bias", "Upper LoA")
+
+  if(sum(df_glm$failure) == 0){
+    message("All values within delta bounds. Confidence intervals hypothesis test likely bogus.")
+    test = binom.test(sum(df_glm$success), nrow(df_glm))
+    df_agree = data.frame(
+      row.names = paste0("Prop. within ",dbound),
+      agreement = test$estimate,
+      lower.ci = test$conf.int[1],
+      upper.ci = test$conf.int[2]
+    )
+  } else{
+
+  glm_mod = glm(cbind(success,failure) ~ 1,
+                data = df_glm,
+                family = binomial)
+  glm_emm = as.data.frame(confint(emmeans(glm_mod, ~1,
+                                          type = "response"),
+                                  level = conf.level))
+  df_agree = data.frame(agreement = glm_emm$prob,
[…]
