+++ NAMESPACE
@@ -10,6 +10,7 @@ S3method(print,loa_mixed_bs)
 S3method(print,simple_agree)
 S3method(print,simple_reli)
 export(agree_nest)
+export(agree_np)
 export(agree_reps)
 export(agree_test)
 export(blandPowerCurve)
@@ -25,6 +26,7 @@ import(dplyr)
 import(emmeans)
 import(ggplot2)
 import(lme4)
+import(quantreg)
 importFrom(dplyr,mutate)
 importFrom(dplyr,rename)
 importFrom(dplyr,select)

[…]
