### 2022-03-20 5c33d5f
prop_bias
proportional bias models included for all simple_agree functions
