+++ R/agree_nest.R
@@ -7,6 +7,8 @@
 #' @param conf.level the confidence level required. Default is 95\%.
 #' @param agree.level the agreement level required. Default is 95\%. The proportion of data that should lie between the thresholds, for 95\% limits of agreement this should be 0.95.
 #' @param delta The threshold below which methods agree/can be considered equivalent, can be in any units. Equivalence Bound for Agreement.
+#' @param prop_bias Logical indicator (TRUE/FALSE) of whether proportional bias should be considered for the limits of agreement calculations.
+#' @param TOST Logical indicator (TRUE/FALSE) of whether to use two one-tailed tests for the limits of agreement. Default is TRUE.
 #'
 #' @return Returns single simple_agree class object with the results of the agreement analysis.
 #'
@@ -44,14 +46,23 @@ agree_nest <- function(x,
                        data,
                        delta,
                        agree.level = .95,
-                       conf.level = .95){
+                       conf.level = .95,
+                       TOST = TRUE,
+                       prop_bias = FALSE){
 
   agreeq = qnorm(1 - (1 - agree.level) / 2)
   agree.l = 1 - (1 - agree.level) / 2
   agree.u = (1 - agree.level) / 2
   confq = qnorm(1 - (1 - conf.level) / 2)
-  alpha.l = 1 - (1 - conf.level) / 2
-  alpha.u = (1 - conf.level) / 2
+  if(TOST == TRUE){
+    confq2 = qnorm(1 - (1 - conf.level) )
+    alpha.l = 1 - (1 - conf.level)
+    alpha.u = (1 - conf.level)
+  } else {
+    confq2 = qnorm(1 - (1 - conf.level) / 2)
+    alpha.l = 1 - (1 - conf.level) / 2
+    alpha.u = (1 - conf.level) / 2
+  }
 
   df = data %>%
     select(all_of(id),all_of(x),all_of(y)) %>%
@@ -75,7 +86,9 @@ agree_nest <- function(x,
                       lower.ci = ccc_nest[2],
                       upper.ci = ccc_nest[3],
                       SE = ccc_nest[4])
-
+  df_lmer = df %>%
+    mutate(mean = (x+y)/2,
+           delta = x - y)
   df2 = df %>%
     group_by(id) %>%
     summarize(m = n(),
@@ -90,13 +103,40 @@ agree_nest <- function(x,
 
   df3 = df2 %>%
     drop_na()
+  d_varl = c()
+  if(prop_bias == TRUE){
+    form1 = as.formula(delta ~ mean)
+  } else{
+    form1 = as.formula(delta ~ 1)
+  }
+  # Get variance per id
+  for(i in 1:nrow(df3)){
+    idtemp = df3[i,]$id
+    dftemp = df %>% filter(id == idtemp) %>%
+      mutate(mean = (x+y)/2,
+             delta = x - y)
+    d_varl[i] = sigma(lm(delta ~ form1,
+                         data = dftemp))^2
+  }
 
-  d_bar = mean(df2$d)
-  d_var = var(df2$d)
-  sdw2 = sum((df3$m-1)/(nrow(df)-nrow(df3))*df3$d_var)
+  if(prop_bias == FALSE){
+    d_bar = mean(df2$d)
+    d_var = var(df2$d)
+    d_lo = d_bar - confq*sqrt(d_var)/sqrt(nrow(df2))
+    d_hi = d_bar + confq*sqrt(d_var)/sqrt(nrow(df2))
+  } else{
[…]

[…]
