+++ R/agree_nest.R
@@ -58,10 +58,12 @@ agree_nest <- function(x,
     confq2 = qnorm(1 - (1 - conf.level) )
     alpha.l = 1 - (1 - conf.level)
     alpha.u = (1 - conf.level)
+    conf2 = 1 - (1 - conf.level) * 2
   } else {
     confq2 = qnorm(1 - (1 - conf.level) / 2)
     alpha.l = 1 - (1 - conf.level) / 2
     alpha.u = (1 - conf.level) / 2
+    conf2 = conf.level
   }
 
   df = data %>%
@@ -115,7 +117,7 @@ agree_nest <- function(x,
     dftemp = df %>% filter(id == idtemp) %>%
       mutate(mean = (x+y)/2,
              delta = x - y)
-    d_varl[i] = sigma(lm(delta ~ form1,
+    d_varl[i] = sigma(lm(form1,
                          data = dftemp))^2
   }
 
@@ -161,10 +163,14 @@ agree_nest <- function(x,
   loa_u.l = loa_u - RME
   loa_u.u = loa_u + LME
 
+  if(prop_bias == TRUE){
+    message("prop_bias set to TRUE. Hypothesis test may be bogus. Check plots.")
+  }
   df_loa = data.frame(
     estimate = c(d_bar, loa_l, loa_u),
     lower.ci = c(d_lo, loa_l.l, loa_u.l),
     upper.ci = c(d_hi, loa_l.u, loa_u.u),
+    ci.level = c(conf.level, conf2, conf2),
     row.names = c("Bias","Lower LoA","Upper LoA")
   )
   if (!missing(delta)) {
@@ -191,6 +197,12 @@ agree_nest <- function(x,
 
   if(is.null(call2$conf.level)){
     call2$conf.level = conf.level
+  }
+  if(is.null(call2$TOST)){
+    call2$TOST = TOST
+  }
+  if(is.null(call2$prop_bias)){
+    call2$prop_bias = prop_bias
   }
   call2$lm_mod = lm_mod
 
