+++ R/methods.simple_agree.R
@@ -24,35 +24,43 @@
 #' @export
 
 print.simple_agree <- function(x,...){
-  if(x$call == "agree_test") {
+  conf_TOST = if(x$call$TOST){
+    1-(1-x$call$conf.level)*2
+  } else {
+    x$call$conf.level
+  }
+  if(as.character(x$call[1]) == "agree_test") {
   cat("Limit of Agreement = ", x$shieh_test$prop0*100, "%",  sep = "")
+  #cat("\n")
+  #cat("alpha =", (1-x$call$conf.level), "|", x$call$conf.level*100,"% Confidence Interval")
   cat("\n")
-  cat("alpha =", (1-x$call$conf.level), "|", x$call$conf.level*100,"% Confidence Interval")
   cat("\n")
+  cat("###- Shieh Results -###")
   cat("\n")
-  cat("###- Shieh TOST Results -###")
-  cat("\n")
-  cat("Exact C.I.:"," [",round(x$shieh_test$lower.ci,4),", ",round(x$shieh_test$upper.ci, 4), "]", sep = "")
+  cat("Exact ",100*conf_TOST,"% C.I. "," [",round(x$shieh_test$lower.ci,4),", ",round(x$shieh_test$upper.ci, 4), "]", sep = "")
   cat("\n")
   cat("Hypothesis Test: ",x$shieh_test$h0_test, sep = "")
   cat("\n")
+  if(x$call$prop_bias == TRUE) {cat("Warning: hypothesis test likely bogus with proportional bias.")}
   cat("\n")
   cat("###- Bland-Altman Limits of Agreement (LoA) -###")
   cat("\n")
   print(x$loa, digits = 4)
   cat("\n")
-  cat("\n")
+  if(x$call$prop_bias == TRUE) {cat("LoA at average of both measures. Please check plot.")
+    cat("\n")}
   cat("###- Concordance Correlation Coefficient (CCC) -###")
   cat("\n")
   cat("CCC: ",round(x$ccc.xy$est.ccc,4),", ",100*x$call$conf.level,"% C.I. ","[",round(x$ccc.xy$lower.ci,4),", ",round(x$ccc.xy$upper.ci,4),"]",sep = "")
   cat("\n")
-  } else if(x$call[1] == "agree_reps()"){
+  } else if(as.character(x$call[1]) == "agree_reps"){
     cat("Limit of Agreement = ", x$call$conf.level*100, "%",  sep = "")
-    cat("\n")
-    cat("alpha =", (1-x$call$conf.level), "|", x$call$conf.level*100,"% Confidence Interval")
+    #cat("\n")
+    #cat("alpha =", (1-x$call$conf.level), "|", x$call$conf.level*100,"% Confidence Interval")
     cat("\n")
     cat("Replicate Data Points (true value does not vary)")
     cat("\n")
+    if(x$call$prop_bias == TRUE) {cat("Warning: hypothesis test likely bogus with proportional bias.")}
     cat("\n")
     cat("Hypothesis Test: ",x$h0_test, sep = "")
     cat("\n")
@@ -61,20 +69,22 @@ print.simple_agree <- function(x,...){
     cat("\n")
     print(x$loa, digits = 4)
     cat("\n")
-    cat("\n")
+    if(x$call$prop_bias == TRUE) {cat("LoA at average of both measures. Please check plot.")
+      cat("\n")}
     cat("###- Concordance Correlation Coefficient* (CCC) -###")
     cat("\n")
[…]
