+++ R/globals.R
@@ -21,4 +21,13 @@ utils::globalVariables(c("boot_bias",
                          "predicted",
                          "the_int",
                          "the_slope",
-                         "ucl", 'y'))
+                         "ucl", 'y',
+                         "mean_y", "n_y",
+                         "diff_y2", "n_x",
+                         "diff_x", "mean_x",
+                         "diff_y", "mean_y",
+                         "at", "median", "confint",
+                         "binomial", "glm",
+                         "binom.test", "failure",
+                         "success",
+                         "diff_x2"))
