### 2022-03-21 #4148 @arcaldwell49
@Kevin-Mattheus-Moerman  Apologies, I thought I posted this here, but the paper has been updated to address those comments

https://raw.githubusercontent.com/arcaldwell49/SimplyAgree/master/paper/paper.md
