### 2022-03-21 cccb28c
Deming regression function
Added basics
