### 2022-03-21 d9a9e89
minimally working deming reg
not working for replicates yet...
