+++ NAMESPACE
@@ -5,9 +5,11 @@ S3method(check,simple_reli)
 S3method(plot,loa_mixed_bs)
 S3method(plot,powerCurve)
 S3method(plot,simple_agree)
+S3method(plot,simple_deming)
 S3method(plot,simple_reli)
 S3method(print,loa_mixed_bs)
 S3method(print,simple_agree)
+S3method(print,simple_deming)
 S3method(print,simple_reli)
 export(agree_nest)
 export(agree_np)
@@ -15,6 +17,7 @@ export(agree_reps)
 export(agree_test)
 export(blandPowerCurve)
 export(check)
+export(dem_reg)
 export(find_n)
 export(jmvagree)
 export(jmvagreemulti)

[…]
