### 2022-03-22 #4148 @xuanxu
> I think we have an issue with multiple affiliations for one author. Here the paper features the yaml `affiliation: "1, 2"` However, the affiliation appears as superscript `12` rather than `1, 2`.

Fix deployed. This should be `1, 2` now.
